+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 [Unreleased]: https://github.com/atomist/k8-automation/compare/0.1.1...HEAD
 
+### Changed
+
+-  Use nginx-ingress rather than default for GKE
+
 ## [0.1.1][] - 2018-02-22
 
 [0.1.1]: https://github.com/atomist/k8-automation/compare/0.1.0...0.1.1