+++ CHANGELOG.md
@@ -7,7 +7,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased][]
 
-[Unreleased]: https://github.com/atomist/k8-automation/compare/0.1.1...HEAD
+[Unreleased]: https://github.com/atomist/k8-automation/compare/0.2.0...HEAD
+
+## [0.2.0][] - 2018-02-23
+
+[0.2.0]: https://github.com/atomist/k8-automation/compare/0.1.1...0.2.0
+
+Ingress release
 
 ### Changed